Add rudimentary support for use on OS X.
Change spotify-debus-call into macro so we can decide if we need to
call applescript or dbus in compile-time.

Change name of spotify-debus-call to spotify-call to reflect the fact
that it now can be called on OS X as well.

Decide on eval if defuns and other things are needed by checking for
platform.